+++ package.json
@@ -35,8 +35,6 @@
     "react-native-vector-icons": "^10.2.0",
     "react-native-zip-archive": "^7.0.1",
     "socket.io-client": "^4.8.1",
-    "wad": "file:",
-    "WAD": "file:",
     "WADApp": "file:"
   },
   "devDependencies": {
